mux API を整理して公開導線を集約する
